commit: Refactor ProTable component and add Sortable component for sorting functionality
In this commit, I refactored the ProTable component by adding a new prop 'sort' to accept an array of strings. Additionally, I created a new Sortable component that handles the sorting logic and sends the updated sort state back to the parent ProTable component via a callback function passed as a prop. The sorting functionality is now implemented using the Material-UI TableSortIcon and ActionIcon components from @mantine/core and @tabler/icons-react, respectively.

Furthermore, I made changes to the medication index page by importing and using the Sortable component in the ProTable headerSection. The sort state is now managed using React's useState hook and passed down as a prop to the Sortable component. When a sort icon is clicked, the sort state is updated accordingly, and the onChange callback function is called with the new sort value.

Lastly, I made adjustments to the medicationApi service by adding an optional params object to the query function definition. This change allows for passing sorting information as query parameters when making API requests.