refactor: remove unused imports and dependencies in classification-disease.controller.ts and diagnose.controller.ts
In the first file, I removed the imports for `ParseBoolPipe`, `PaginationQuery`, `paginate` and `Pagination` since they are not being used in the controller. Also, I imported `EntityManager` directly instead of importing it and `IsNull` and `Not` from `typeorm`.

In the second file, I added an import for `Payment` class and removed the unused imports for `EncounterPayment`, `PaymentStatus`, `JwtAuthGuard`, `PaymentMethd`, and `DonePaymentInputRequest`.

Additionally, in the payment.controller.ts file, I added `@UseGuards(JwtAuthGuard)` decorator to some endpoints and imported `FilterOperator`, `PaginateQuery`, `paginate` from 'nestjs-paginate' package.