In the first commit, I changed the port variable name from lowercase 'port' to uppercase 'PORT' in the server file for better semantics and readability.
In the second commit, I added new options "Supplement" and "Sol" with their respective display names and definitions in the dose-form.data.json file.

In the third commit, I changed the sort type from lowercase 'sort' to uppercase 'Sort' in the Sortable component for better semantics and readability. Additionally, I updated the use of sort in the Component page by passing it as a prop to the PaginateQuery hook instead of using it directly as a state.